requirements.txt
requests==2.26.0
paho-mqtt==1.5.1
pydantic==1.8.2
PyYAML==5.4.1